Redux integration (#146)
* redux is added to everything but livelocation

* redux is working

* formatted code
