+++ c2tc-mobile/components/Loader.js
@@ -5,7 +5,14 @@ const Loader = props => {
   const { loading, ...attributes } = props;
 
   return (
-    <Modal transparent={true} animationType={"none"} visible={loading}>
+    <Modal
+      transparent={true}
+      onRequestClose={() => {
+        Alert.alert("Modal has been closed.");
+      }}
+      animationType={"none"}
+      visible={loading}
+    >
       <View style={styles.modalBackground}>
         <View style={styles.activityIndicatorWrapper}>
           <ActivityIndicator animating={loading} />
