+++ c2tc-mobile/package.json
@@ -38,6 +38,8 @@
     "react-native-sliding-up-down-panels": "^1.0.0",
     "react-native-tabs": "^1.0.9",
     "react-navigation": "^2.18.3",
+    "react-redux": "^6.0.1",
+    "redux": "^4.0.1",
     "rn-sliding-up-panel": "^1.3.1"
   },
   "devDependencies": {
