+++ c2tc-mobile/screens/LiveLocation.js
@@ -1,15 +1,22 @@
 import React, { Component } from "react";
 import { StyleSheet, View, Dimensions, AsyncStorage } from "react-native";
 import { Location, Permissions } from "expo";
-
 import MapView, { Marker, ProviderPropType } from "react-native-maps";
 import Navigation from "../components/NavigationComponents/Navigation";
 import Colors from "../constants/Colors";
-
+import renderLayerMarkers from "../components/MapRendering";
 import API from "../components/API";
 import Loader from "../components/Loader";
 
 import CurrentLocationButton from "../components/NavigationComponents/CurrentLocationButton";
+import { connect } from "react-redux";
+import { bindActionCreators } from "redux";
+import {
+  updateMapRegion,
+  updateColorData,
+  updateLayerData,
+  updateDetailView
+} from "../Redux";
 
 const { width, height } = Dimensions.get("window");
 
@@ -17,15 +24,26 @@ const ASPECT_RATIO = width / height;
 const LATITUDE_DELTA = 0.017;
 const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
 
-let id = 0;
-
-const icons = {
-  busStop: require("../assets/images/bus.png"),
-  crime: require("../assets/images/crime.png"),
-  business: require("../assets/images/business.png"),
-  emergency: require("../assets/images/phone.png"),
-  policeStations: require("../assets/images/police.png"),
-  streetLights: require("../assets/images/streetlights.png")
+const mapDispatchToProps = dispatch => {
+  return bindActionCreators(
+    {
+      updateColorData,
+      updateLayerData,
+      updateMapRegion,
+      updateDetailView
+    },
+    dispatch
+  );
+};
+const mapStateToProps = state => {
+  return {
+    markerClicked: state.markerClicked,
+    layerData: state.layerData,
+    colorData: state.colorData,
+    renderData: state.renderData,
+    markers: state.markers,
+    mapRegion: state.mapRegion
+  };
 };
 
 class LiveLocation extends Component {
@@ -33,22 +51,8 @@ class LiveLocation extends Component {
     super(props);
 
     this.state = {
-      mapRegion: null,
       lastLat: null,
       lastLong: null,
-      markers: [],
-      renderData: {
-        busStop: false,
-        crime: false,
-        business: false,
-        emergency: false
-      },
-      layerData: {},
-      loading: true,
-      colorData: {},
-      markerClicked: false,
-      markerTitle: "",
-      markerDescrption: "",
       locationResult: null
     };
   }
@@ -62,15 +66,6 @@ class LiveLocation extends Component {
   async componentDidMount() {
     this._mounted = true;
 
-    this.watchID = navigator.geolocation.watchPosition(position => {
-      let region = {
-        latitude: position.coords.latitude,
-        longitude: position.coords.longitude,
-        latitudeDelta: LATITUDE_DELTA,
-        longitudeDelta: LONGITUDE_DELTA
-      };
-      this.onRegionChange(region, region.latitude, region.longitude);
-    });
     navigator.geolocation.getCurrentPosition(
       position => {
         let region = {
@@ -79,12 +74,12 @@ class LiveLocation extends Component {
           latitudeDelta: LATITUDE_DELTA,
           longitudeDelta: LONGITUDE_DELTA
         };
-        this.onRegionChange(region, region.latitude, region.longitude);
+        this.onRegionChange(region);
       },
       error => console.log({ error: error.message })
     );
 
-    this.watchID = await navigator.geolocation.watchPosition(
+    this.watchID = navigator.geolocation.watchPosition(
       position => {
         let region = {
           latitude: position.coords.latitude,
@@ -92,20 +87,12 @@ class LiveLocation extends Component {
           latitudeDelta: LATITUDE_DELTA,
           longitudeDelta: LONGITUDE_DELTA
         };
-        this.onRegionChange(region, region.latitude, region.longitude);
+        this.onRegionChange(region);
       },
       error => console.log({ error: error.message })
     );
 
-    for (var layer in this.state.layerData) {
-      this.renderMarkers(
-        layer,
-        this.state.layerData[layer],
-        this.state.colorData[layer]
-      );
-    }
-
-    if (AsyncStorage.getAllKeys().length != 4) {
+    if (AsyncStorage.getAllKeys().length != 6) {
       let busStopData = await API.getBusStops();
       let crimeData = await API.getCrimes();
       let businessData = await API.getBusinesses();
@@ -126,253 +113,91 @@ class LiveLocation extends Component {
       );
       await AsyncStorage.setItem("streetLights", JSON.stringify(streetLights));
     }
+    await this.props.updateLayerData({
+      busStop: JSON.parse(await AsyncStorage.getItem("busStop")),
+      crime: JSON.parse(await AsyncStorage.getItem("crimeData")),
+      business: JSON.parse(await AsyncStorage.getItem("businessData")),
+      emergency: JSON.parse(await AsyncStorage.getItem("emergencyData")),
+      policeStations: JSON.parse(await AsyncStorage.getItem("policeStations")),
+      streetLights: JSON.parse(await AsyncStorage.getItem("streetLights"))
+    });
 
-    if (this._mounted) {
-      this.setState({
-        layerData: {
-          busStop: JSON.parse(await AsyncStorage.getItem("busStop")),
-          crime: JSON.parse(await AsyncStorage.getItem("crimeData")),
-          business: JSON.parse(await AsyncStorage.getItem("businessData")),
-          emergency: JSON.parse(await AsyncStorage.getItem("emergencyData")),
-          policeStations: JSON.parse(
-            await AsyncStorage.getItem("policeStations")
-          ),
-          streetLights: JSON.parse(await AsyncStorage.getItem("streetLights"))
-        },
-        colorData: {
-          busStop: Colors.busStop,
-          crime: Colors.crime,
-          business: Colors.business,
-          emergency: Colors.emergency,
-          policeStations: Colors.police,
-          streetLights: Colors.streetlights
-        },
-        loading: false
-      });
-    }
+    await this.props.updateColorData({
+      busStop: Colors.busStop,
+      crime: Colors.crime,
+      business: Colors.business,
+      emergency: Colors.emergency,
+      policeStations: Colors.police,
+      streetLights: Colors.streetlights
+    });
 
+    this._mounted = false;
     this.getLocationAsync();
   }
 
-  onRegionChange(region, lastLat, lastLong) {
-    if (this._mounted) {
-      this.setState({
-        mapRegion: region,
-        lastLat: lastLat || this.state.lastLat,
-        lastLong: lastLong || this.state.lastLong
-      });
-    }
-  }
+  onRegionChange = region => {
+    this.props.updateMapRegion(region);
+    this.setState({
+      lastLat: region.latitude || this.state.lastLat,
+      lastLong: region.longitude || this.state.lastLong
+    });
+  };
 
   componentWillUnmount() {
     navigator.geolocation.clearWatch(this.watchID);
-    this._mounted = false;
-  }
-
-  getDistance(lat1, lon1, lat2, lon2) {
-    let earthRadius = 6371;
-    let deltaLat = this.toRad(lat2 - lat1);
-    let deltaLong = this.toRad(lon2 - lon1);
-    let currentLat = this.toRad(lat1);
-    let finalLat = this.toRad(lat2);
-
-    let pythag =
-      Math.sin(deltaLat / 2) * Math.sin(deltaLat / 2) +
-      Math.sin(deltaLong / 2) *
-        Math.sin(deltaLong / 2) *
-        Math.cos(currentLat) *
-        Math.cos(finalLat);
-    let deriv = 2 * Math.atan2(Math.sqrt(pythag), Math.sqrt(1 - pythag));
-    let mult = earthRadius * deriv;
-    kmToMiles = mult / 1.6;
-    return Math.round(kmToMiles * 100) / 100;
-  }
-
-  toRad(value) {
-    return (value * Math.PI) / 180;
   }
 
   componentWillMount() {
     setTimeout(() => {
-      if (this._mounted) {
-        this.setState({ statusBarHeight: 5 });
-      }
+      this.setState({ statusBarHeight: 5 });
     }, 500);
   }
 
-  renderMarkers(layer, data, markerColor) {
-    data = this.state.layerData[layer];
-    let list = this.state.markers;
-    for (i = 0; i < data.length; i++) {
-      if (markerColor === this.state.colorData.busStop) {
-        buses = "";
-        for (let key in data[i].routes) {
-          if (key == data[i].routes[data[i].routes.length - 1]) {
-            buses = buses + key + ".";
-          } else {
-            buses = buses + key + ", ";
-          }
-        }
-        title = data[i].stop_name;
-        description = "Buses come to this stop: " + buses;
-      } else if (markerColor === this.state.colorData.emergency) {
-        distance = this.getDistance(
-          data[i].latitude,
-          data[i].longitude,
-          this.state.mapRegion.latitude,
-          this.state.mapRegion.longitude
-        );
-        title = distance + " miles away";
-        description = "Emergency Phone #" + data[i].emergencyPhone_id;
-      } else if (markerColor === this.state.colorData.crime) {
-        distance = this.getDistance(
-          data[i].latitude,
-          data[i].longitude,
-          this.state.mapRegion.latitude,
-          this.state.mapRegion.longitude
-        );
-        title = data[i].incident_type_primary;
-        description =
-          data[i].incident_datetime +
-          "\n" +
-          distance +
-          " miles away \n" +
-          data[i].incident_description;
-      } else if (markerColor === this.state.colorData.business) {
-        address = "";
-        for (let key in data[i].location) {
-          if (
-            data[i].location[key] ==
-            data[i].location[data[i].location.length - 1]
-          ) {
-            address = address + data[i].location[key] + ".";
-          } else {
-            address = address + data[i].location[key] + ", ";
-          }
-        }
-        title = data[i].name;
-        description = "Address: " + address;
-      } else if (markerColor === this.state.colorData.policeStations) {
-        title = data[i].name;
-        description = data[i].name + " is located here";
-      } else if (markerColor === this.state.colorData.streetLights) {
-        title = "Streetlight";
-        description = "";
-      } else {
-        title = "Title";
-        description = "Description";
-      }
-      list.push({
-        coordinate: {
-          latitude: data[i].latitude,
-          longitude: data[i].longitude
-        },
-        key: id++,
-        color: markerColor,
-        image: icons[layer],
-        title: title,
-        description: description
-      });
-    }
-    if (this._mounted) {
-      this.setState({
-        markers: list
-      });
-    }
-  }
-
   markerClick = (title, description) => {
-    if (this._mounted) {
-      this.setState({
-        markerClicked: true,
-        markerTitle: title,
-        markerDescrption: description
-      });
-    }
-  };
-
-  changeMarkerToFalse = () => {
-    if (this._mounted) {
-      this.setState({
-        markerClicked: false
-      });
-    }
-  };
-
-  _onPressToggleLayers = layer => {
-    if (this.state.renderData[layer]) {
-      if (this._mounted) {
-        this.setState({
-          markers: this.state.markers.filter(
-            marker => marker["color"] !== this.state.colorData[layer]
-          )
-        });
-        this.state.renderData[layer] = false;
-      }
-    } else {
-      this.renderMarkers(
-        layer,
-        this.state.layerData[layer],
-        this.state.colorData[layer]
-      );
-      if (this._mounted) {
-        this.state.renderData[layer] = true;
-      }
-    }
+    this.props.updateDetailView(true, title, description);
   };
 
   backToUser = () => {
-    if (this._mounted) {
-      this.setState({
-        mapRegion: this.state.locationResult
-      });
-    }
+    this.props.updateMapRegion(this.state.locationResult);
   };
 
   onRegionChangeRender = region => {
-    this.state.mapRegion = region;
+    this.props.updateMapRegion = region;
   };
 
   getLocationAsync = async () => {
     let { status } = await Permissions.askAsync(Permissions.LOCATION);
     if (status !== "granted") {
-      if (this._mounted) {
-        this.setState({
-          locationResult: "Permission to access location was denied"
-        });
-      }
+      this.setState({
+        locationResult: "Permission to access location was denied"
+      });
     }
 
     let location = await Location.getCurrentPositionAsync({});
-    let locationTwo = {
+    let region = {
       latitude: location.coords.latitude,
-      latitudeDelta: LATITUDE_DELTA,
       longitude: location.coords.longitude,
+      latitudeDelta: LATITUDE_DELTA,
       longitudeDelta: LONGITUDE_DELTA
     };
-    if (this._mounted) {
-      this.setState({ locationResult: locationTwo });
-    }
+    this.setState({ locationResult: region });
   };
 
   render() {
     if (this._mounted) {
-      if (this.state.loading) {
-        return <Loader loading={this.state.loading} />;
-      }
+      return <Loader loading={this._mounted} />;
     }
-
     return (
       <View style={styles.container}>
         <MapView
           style={styles.map}
-          region={this.state.mapRegion}
+          region={this.props.mapRegion}
           showsUserLocation={true}
           followUserLocation={true}
           showsMyLocationButton={true}
           onRegionChange={this.onRegionChangeRender}
         >
-          {this.state.markers.map(marker => (
+          {this.props.markers.map(marker => (
             <Marker
               key={marker.key}
               coordinate={marker.coordinate}
@@ -392,15 +217,7 @@ class LiveLocation extends Component {
           <CurrentLocationButton changeLocation={this.backToUser} />
         </View>
 
-        <Navigation
-          ref="panel"
-          description={this.state.markerClicked}
-          descriptionTitle={this.state.markerTitle}
-          descriptionContent={this.state.markerDescrption}
-          onDescExit={this.changeMarkerToFalse}
-          toggleLayers={this._onPressToggleLayers}
-          layers={this.state.renderData}
-        />
+        <Navigation ref="panel" layers={this.props.renderData} />
       </View>
     );
   }
@@ -409,6 +226,11 @@ class LiveLocation extends Component {
 LiveLocation.propTypes = {
   provider: ProviderPropType
 };
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(LiveLocation);
 
 const styles = StyleSheet.create({
   zoom: {
@@ -447,5 +269,3 @@ const styles = StyleSheet.create({
     backgroundColor: "transparent"
   }
 });
-
-export default LiveLocation;
